update to libarchive 3.5.2

+++ qlZipInfo/libarchive/archive_getdate.c
@@ -960,7 +960,7 @@ difftm (struct tm *a, struct tm *b)
 {
 	int ay = a->tm_year + (TM_YEAR_ORIGIN - 1);
 	int by = b->tm_year + (TM_YEAR_ORIGIN - 1);
-	int days = (int)(
+	int days = (
 		/* difference in day of year */
 		a->tm_yday - b->tm_yday
 		/* + intervening leap days */
@@ -968,7 +968,7 @@ difftm (struct tm *a, struct tm *b)
 		-  (ay/100 - by/100)
 		+  ((ay/100 >> 2) - (by/100 >> 2))
 		/* + difference in years * 365 */
-		+  (long)(ay-by) * 365
+		+  (int)((long)(ay-by) * 365)
 		);
 	return (days * DAY + (a->tm_hour - b->tm_hour) * HOUR
 	    + (a->tm_min - b->tm_min) * MINUTE
